resources: move from src to lib folder, fix Api usage.
