+++ lib/resources/project-data.js
@@ -6,19 +6,19 @@
  * LICENSE file in the root directory.
  */
 
-const ApiRequest = require('../request/api')
+const Api = require('../api')
 
-const ProjectDataClass = require('../../classes/project-data')
+const ProjectDataClass = require('../project-data')
 
-const Params = require('../../util/params')
+const Params = require('../util/params')
 
-const { PROJECT_DATA_API_PREFIX } = require('../../config')
+const { PROJECT_DATA_API_PREFIX } = require('../config')
 
 class ProjectData {
   /**
+   * Get Project-data.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Get Project-data.
    */
   static getProjectData (payload) {
     const projectId = Params.destructURLParam(payload, 'projectId')
@@ -26,14 +26,14 @@ class ProjectData {
     const options = {
       endpoint: `${PROJECT_DATA_API_PREFIX}/project-data/${projectId}`
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
       .then((projectDataJson) => new ProjectDataClass(projectDataJson))
   }
 
   /**
+   * Update Project-data (partial update)
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Update Project-data (partial update)
    */
   static updateProjectDataPartial (payload) {
     const body = Params.destructParams(payload, ['data'])
@@ -44,7 +44,7 @@ class ProjectData {
       endpoint: `${PROJECT_DATA_API_PREFIX}/project-data/${projectId}`,
       body
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 }
