+++ lib/resources/projects.js
@@ -6,17 +6,17 @@
  * LICENSE file in the root directory.
  */
 
-const ApiRequest = require('../request/api')
+const Api = require('../api')
 
-const Params = require('../../util/params')
+const Params = require('../util/params')
 
-const { API_PREFIX } = require('../../config')
+const { API_PREFIX } = require('../config')
 
 class Projects {
   /**
+   * Get All Projects.
    * @param {Object} [payload]
    * @returns {Promise.<Array>}
-   * @description Get All Projects.
    */
   static getProjects (payload) {
     const qs = Params.destructParams(payload,
@@ -26,13 +26,13 @@ class Projects {
       endpoint: `${API_PREFIX}/projects`,
       qs
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Add Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Add Project.
    */
   static addProject (payload) {
     const body = Params.destructParams(payload, ['templateId'])
@@ -42,13 +42,13 @@ class Projects {
       endpoint: `${API_PREFIX}/projects`,
       body
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Get Trial Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Get Trial Project.
    */
   static getTrialProject (payload) {
     const qs = Params.destructParams(payload, ['templateId'])
@@ -57,13 +57,13 @@ class Projects {
       endpoint: `${API_PREFIX}/projects/trial`,
       qs
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get a Specific Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Get a Specific Project.
    */
   static getProject (payload) {
     const projectId = Params.destructURLParam(payload, 'projectId')
@@ -71,13 +71,13 @@ class Projects {
     const options = {
       endpoint: `${API_PREFIX}/projects/${projectId}`
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Update the Project (partial update).
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Update the Project (partial update).
    */
   static updateProjectPartial (payload) {
     const body = Params.destructParams(payload, ['customTitle'])
@@ -88,13 +88,13 @@ class Projects {
       endpoint: `${API_PREFIX}/projects/${projectId}`,
       body
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Delete a Specific Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Delete a Specific Project.
    */
   static deleteProject (payload) {
     const projectId = Params.destructURLParam(payload, 'projectId')
@@ -103,13 +103,13 @@ class Projects {
       method: 'DELETE',
       endpoint: `${API_PREFIX}/projects/${projectId}`
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Delete Specific Project Videos.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Delete Specific Project Videos.
    */
   static deleteProjectVideos (payload) {
     const projectId = Params.destructURLParam(payload, 'projectId')
@@ -119,13 +119,13 @@ class Projects {
       method: 'DELETE',
       endpoint: `${API_PREFIX}/projects/${projectId}/videos/${quality}`
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Apply Template Preset on the Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Apply Template Preset on the Project.
    */
   static applyTemplatePresetOnProject (payload) {
     const body = Params.destructParams(payload, ['presetId'])
@@ -136,13 +136,13 @@ class Projects {
       endpoint: `${API_PREFIX}/projects/${projectId}/apply-template-preset`,
       body
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Duplicate the Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Duplicate the Project.
    */
   static duplicateProject (payload) {
     const projectId = Params.destructURLParam(payload, 'projectId')
@@ -151,13 +151,13 @@ class Projects {
       method: 'POST',
       endpoint: `${API_PREFIX}/projects/${projectId}/duplicate`
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Render the Project.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Render the Project.
    */
   static renderProject (payload) {
     const body = Params.destructParams(payload, ['quality', 'watermark'])
@@ -168,7 +168,7 @@ class Projects {
       endpoint: `${API_PREFIX}/projects/${projectId}/render`,
       body
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 }
