+++ lib/resources/sounds.js
@@ -6,17 +6,17 @@
  * LICENSE file in the root directory.
  */
 
-const ApiRequest = require('../request/api')
+const Api = require('../api')
 
-const Params = require('../../util/params')
+const Params = require('../util/params')
 
-const { API_PREFIX } = require('../../config')
+const { API_PREFIX } = require('../config')
 
 class Sounds {
   /**
+   * Get Company Sounds (limited).
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Company Sounds (limited).
    */
   static getCompanySoundsLimited (payload) {
     const qs = Params.destructParams(payload, ['duration'])
@@ -25,13 +25,13 @@ class Sounds {
       endpoint: `${API_PREFIX}/sounds`,
       qs
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Sounds.
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Sounds.
    */
   static getSounds (payload) {
     const qs = Params.destructParams(payload, ['duration'])
@@ -40,13 +40,13 @@ class Sounds {
       endpoint: `${API_PREFIX}/sounds`,
       qs
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 
   /**
+   * Get Recommended Sounds (limited).
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Recommended Sounds (limited).
    */
   static getRecommendedSoundsLimited (payload) {
     const qs = Params.destructParams(payload, ['duration', 'templateId'])
@@ -55,13 +55,13 @@ class Sounds {
       endpoint: `${API_PREFIX}/sounds/recommended`,
       qs
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Recommended Sounds.
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Recommended Sounds.
    */
   static getRecommendedSounds (payload) {
     const qs = Params.destructParams(payload, ['duration', 'templateId'])
@@ -70,7 +70,7 @@ class Sounds {
       endpoint: `${API_PREFIX}/sounds/recommended`,
       qs
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 }
