+++ lib/resources/supports.js
@@ -6,17 +6,17 @@
  * LICENSE file in the root directory.
  */
 
-const ApiRequest = require('../request/api')
+const Api = require('../api')
 
-const Params = require('../../util/params')
+const Params = require('../util/params')
 
-const { API_PREFIX } = require('../../config')
+const { API_PREFIX } = require('../config')
 
 class Supports {
   /**
+   * Add Supports Ticket.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Add Supports Ticket.
    */
   static addSupportsTicket (payload) {
     const body = Params.destructParams(payload, ['message', 'mailType', 'subject'])
@@ -26,7 +26,7 @@ class Supports {
       endpoint: `${API_PREFIX}/supports/ticket`,
       body
     }
-    return ApiRequest.authorizedRequest(options)
+    return Api.authorizedRequest(options)
   }
 }
