+++ lib/resources/templates.js
@@ -6,18 +6,17 @@
  * LICENSE file in the root directory.
  */
 
-const ApiRequest = require('../request/api')
-const WebRequest = require('../request/web')
+const Api = require('../api')
 
-const Params = require('../../util/params')
+const Params = require('../util/params')
 
-const { API_PREFIX, WEB_PREFIX } = require('../../config')
+const { API_PREFIX, WEB_PREFIX } = require('../config')
 
 class Templates {
   /**
+   * Get All Templates.
    * @param {Object} [payload]
    * @returns {Promise.<Array>}
-   * @description Get All Templates.
    */
   static getTemplates (payload) {
     const qs = Params.destructParams(payload, ['categoryId', 'equalizer', 'limit', 'offset'])
@@ -26,13 +25,13 @@ class Templates {
       endpoint: `${API_PREFIX}/templates`,
       qs
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Templates Categories.
    * @param {Object} [payload]
    * @returns {Promise.<Array>}
-   * @description Get Templates Categories.
    */
   static getTemplatesCategories (payload) {
     const qs = Params.destructParams(payload, ['language'])
@@ -41,13 +40,13 @@ class Templates {
       endpoint: `${API_PREFIX}/templates/categories`,
       qs
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get a Specific Template.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Get a Specific Template.
    */
   static getTemplate (payload) {
     const qs = Params.destructParams(payload, ['language'])
@@ -57,13 +56,13 @@ class Templates {
       endpoint: `${WEB_PREFIX}${API_PREFIX}/templates/${templateId}`,
       qs
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Color-Presets of the Template.
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Color-Presets of the Template.
    */
   static getTemplateColorPresets (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -71,13 +70,13 @@ class Templates {
     const options = {
       endpoint: `${API_PREFIX}/templates/${templateId}/color-presets`
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Pluggable-Screens of the Template.
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Pluggable-Screens of the Template.
    */
   static getTemplatePluggableScreens (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -85,13 +84,13 @@ class Templates {
     const options = {
       endpoint: `${API_PREFIX}/templates/${templateId}/pluggable-screens`
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Recommended-Custom-Colors of the Template.
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Recommended-Custom-Colors of the Template.
    */
   static getTemplateRecommendedCustomColors (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -99,13 +98,13 @@ class Templates {
     const options = {
       endpoint: `${API_PREFIX}/templates/${templateId}/recommended-custom-colors`
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Template-Presets of the Template.
    * @param {Object} payload
    * @returns {Promise.<Array>}
-   * @description Get Template-Presets of the Template.
    */
   static getTemplatePresets (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -113,13 +112,13 @@ class Templates {
     const options = {
       endpoint: `${API_PREFIX}/templates/${templateId}/template-presets`
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Template-SVG-Content of the Template.
    * @param {Object} payload
    * @returns {Promise.<>}
-   * @description Get Template-SVG-Content of the Template.
    */
   static getTemplateSVGContent (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -128,13 +127,13 @@ class Templates {
       endpoint: `${WEB_PREFIX}/templates/termplatesvg/${templateId}`
     }
 
-    return WebRequest.request(options)
+    return Api.webRequest(options)
   }
 
   /**
+   * Get Theme of the Template.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Get Theme of the Template.
    */
   static getTemplateTheme (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -142,13 +141,13 @@ class Templates {
     const options = {
       endpoint: `${API_PREFIX}/templates/${templateId}/theme`
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 
   /**
+   * Get Transitions of the Template.
    * @param {Object} payload
    * @returns {Promise.<Object>}
-   * @description Get Transitions of the Template.
    */
   static getTemplateTransitions (payload) {
     const templateId = Params.destructURLParam(payload, 'templateId')
@@ -156,7 +155,7 @@ class Templates {
     const options = {
       endpoint: `${API_PREFIX}/templates/${templateId}/transitions`
     }
-    return ApiRequest.unauthorizedRequest(options)
+    return Api.unauthorizedRequest(options)
   }
 }
