+++ lib/resources/users.js
@@ -6,14 +6,14 @@
  * LICENSE file in the root directory.
  */
 
-const ApiRequest = require('../request/api')
+const ApiRequest = require('../api')
 
-const { API_PREFIX } = require('../../config')
+const { API_PREFIX } = require('../config')
 
 class Users {
   /**
+   * Get Current User.
    * @returns {Promise.<Object>}
-   * @description Get Current User.
    */
   static getCurrentUser () {
     const options = {
